Create main.py
This is a Collatz Conjecture program. What this basically does is that it asks you to enter a whole number. If the number you have entered is an even number, it divides the number by 2 or if the number is an odd number, it multiplies the number by 3 and adds 1 to it. This process continues till you get 1 as the end result. Once you get 1 at the end of the conjecture, the program stops working and the conjecture ends. This conjecture was developed by German mathematician Lothar Collatz in 1937 and ever since has intrigued mathematicians all around the world. This conjecture is also commonly called the 3x+1 conjecture and it is said that the conjecture always ends with the numbers 4, 2, 1.